Change DB filename, and Move Identity model
