+++ OptionsWebsite/App_Start/IdentityConfig.cs
@@ -10,7 +10,7 @@ using Microsoft.AspNet.Identity.EntityFramework;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin;
 using Microsoft.Owin.Security;
-using OptionsWebsite.Models;
+using DiplomaDataModel.Identity;
 
 namespace OptionsWebsite
 {
