+++ OptionsWebsite/Controllers/AccountController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
 using OptionsWebsite.Models;
+using DiplomaDataModel.Identity;
 
 namespace OptionsWebsite.Controllers
 {
