+++ OptionsWebsite/Web.config
@@ -9,7 +9,7 @@
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   </configSections>
   <connectionStrings>
-    <add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-OptionsWebsite-20160211060848.mdf;Initial Catalog=aspnet-OptionsWebsite-20160211060848;Integrated Security=True"
+    <add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\DiplomaOptions.mdf;Initial Catalog=DiplomaOptions;Integrated Security=True"
       providerName="System.Data.SqlClient" />
   </connectionStrings>
   <appSettings>
